Added trick to determine if there's a next page in thread list of forum to enable 'Older' button
The trick is to fetch 1 row extra, after the page. The # of rows fetched
is then used to see whether there's another page left (which is the case
if there's 1 extra row fetched). If not (e.g. there were exactly
pageSize rows left or less), no extra page is left so Older button
should be disabled.